Move Default Trace tabs from Resource Metrics to Overview section (#169)
Default Trace Events and Trace Analysis tabs belong under Overview, not
Resource Metrics. Extracted into new DefaultTraceContent UserControl
following the same pattern as DailySummaryContent and CriticalIssuesContent.
Removed dead column filter popup infrastructure from ResourceMetricsContent.
